// src/@types/type.ts
export interface GetParams {
	params?: any;
	[key: string]: any;
}
export interface PostParams {
	params?: any;
	body?: any;
	[key: string]: any;
}

export interface UploadParams {
	params?: any;
	body?: any;
	file?: any;
	[key: string]: any;
}

export interface GetResponse {
	(d: { status?: any; message?: any; data?: any }): any;
}